requirements.txt
-e .
https://www.djangoproject.com/download/1.7c3/tarball/
django-jinja
django-haystack
celery
celery-haystack
